Complete - task 29: BPEL xml source not available in process viewer http://code.google.com/p/bpel-g/issues/detail?id=29
There's a problem mixing xerces with 1.6. Apparently the version of xerces included in the dependencies was not implementing one of the interfaces needed for dom serialization. The solution was to remove xerces as a dependency. Also refactored the process viewer to avoid having to serialize the dom again since it already had the string.